[Prover] Various small cleanups.
