+++ prover-service/src/deployment_information.rs
@@ -2,8 +2,12 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use aptos_build_info::build_information;
+use aptos_crypto::ed25519::Ed25519PublicKey;
 use aptos_infallible::Mutex;
 use std::{collections::BTreeMap, sync::Arc};
+
+// The key used to store the training wheels verification key in the deployment information
+const TRAINING_WHEELS_VERIFICATION_KEY: &str = "training_wheels_verification_key";
 
 /// A simple struct to hold deployment information as key-value pairs
 #[derive(Clone, Debug)]
@@ -37,4 +41,21 @@ impl Default for DeploymentInformation {
     fn default() -> Self {
         Self::new()
     }
+}
+
+/// Creates and returns the deployment information for the prover service
+pub fn get_deployment_information(
+    training_wheels_verification_key: &Ed25519PublicKey,
+) -> DeploymentInformation {
+    // Create the deployment information
+    let mut deployment_information = DeploymentInformation::new();
+
+    // Insert the training wheels verification key into the deployment information.
+    // This is useful for runtime verification (e.g., to ensure the correct key is being used).
+    deployment_information.extend_deployment_information(
+        TRAINING_WHEELS_VERIFICATION_KEY.into(),
+        training_wheels_verification_key.to_string(),
+    );
+
+    deployment_information
 }
