+++ prover-service/src/metrics.rs
@@ -1,14 +1,27 @@
 // Copyright (c) Aptos Foundation
 
-use aptos_logger::{error, warn};
+use crate::prover_config::ProverServiceConfig;
+use aptos_logger::{error, info, warn};
 use aptos_metrics_core::{
     register_histogram, register_int_counter_vec, Histogram, IntCounterVec, TextEncoder,
 };
 use hyper::header::CONTENT_TYPE;
-use hyper::{Body, Method, StatusCode};
+use hyper::service::{make_service_fn, service_fn};
+use hyper::{Body, Method, Server, StatusCode};
 use once_cell::sync::Lazy;
 use prometheus::{proto::MetricFamily, Encoder};
 use std::convert::Infallible;
+use std::net::SocketAddr;
+use std::sync::Arc;
+
+// Constants for the metrics endpoint and response type
+const METRICS_ENDPOINT: &str = "/metrics";
+const PLAIN_CONTENT_TYPE: &str = "text/plain";
+
+// Useful metric labels for counting all metrics
+const TOTAL_METRIC_BYTES_LABEL: &str = "total_bytes";
+const TOTAL_METRIC_FAMILIES_OVER_2000_LABEL: &str = "families_over_2000";
+const TOTAL_METRICS_LABEL: &str = "total";
 
 // TODO: sanity check and expand these metrics!
 
@@ -30,28 +43,47 @@ pub static REQUEST_QUEUE_TIME_SECS: Lazy<Histogram> = Lazy::new(|| {
     .unwrap()
 });
 
-/// Counter for the number of metrics in various states
-pub static NUM_METRICS: Lazy<IntCounterVec> = Lazy::new(|| {
+/// Counter for the number of prover metrics in various states
+pub static PROVER_NUM_METRICS: Lazy<IntCounterVec> = Lazy::new(|| {
     register_int_counter_vec!(
-        "aptos_metrics",
-        "Number of metrics in certain states",
+        "keyless_prover_num_metrics",
+        "Number of keyless prover metrics in certain states",
         &["type"]
     )
     .unwrap()
 });
 
+// Starts a simple metrics server
+pub fn start_metrics_server(prover_service_config: Arc<ProverServiceConfig>) {
+    let _handle = tokio::spawn(async move {
+        info!("Starting metrics server request handler...");
+
+        // Create a service function that handles the metrics requests
+        let make_service = make_service_fn(|_conn| async {
+            Ok::<_, Infallible>(service_fn(handle_metrics_request))
+        });
+
+        // Bind the socket address, and start the server
+        let socket_addr = SocketAddr::from(([0, 0, 0, 0], prover_service_config.metrics_port));
+        let server = Server::bind(&socket_addr).serve(make_service);
+        if let Err(error) = server.await {
+            panic!("Metrics server error! Error: {}", error);
+        }
+    });
+}
+
 /// Handles incoming HTTP requests for the metrics server
-pub async fn handle_metrics_request(
+async fn handle_metrics_request(
     request: hyper::Request<Body>,
 ) -> Result<hyper::Response<Body>, Infallible> {
     let response = match (request.method(), request.uri().path()) {
-        (&Method::GET, "/metrics") => {
+        (&Method::GET, METRICS_ENDPOINT) => {
             let buffer = get_encoded_metrics(TextEncoder::new());
             hyper::Response::builder()
                 .status(StatusCode::OK)
-                .header(CONTENT_TYPE, "text/plain")
+                .header(CONTENT_TYPE, PLAIN_CONTENT_TYPE)
                 .body(Body::from(buffer))
-                .expect("Metric response should build")
+                .expect("The metric response failed to build!")
         }
         _ => {
             let mut response = hyper::Response::new(Body::empty());
@@ -73,8 +105,8 @@ fn get_encoded_metrics(encoder: impl Encoder) -> Vec<u8> {
     }
 
     // Update the total metric bytes counter
-    NUM_METRICS
-        .with_label_values(&["total_bytes"])
+    PROVER_NUM_METRICS
+        .with_label_values(&[TOTAL_METRIC_BYTES_LABEL])
         .inc_by(encoded_buffer.len() as u64);
 
     encoded_buffer
@@ -104,9 +136,11 @@ fn get_metric_families() -> Vec<MetricFamily> {
     }
 
     // These metrics will be reported on the next pull, rather than create a new family
-    NUM_METRICS.with_label_values(&["total"]).inc_by(total);
-    NUM_METRICS
-        .with_label_values(&["families_over_2000"])
+    PROVER_NUM_METRICS
+        .with_label_values(&[TOTAL_METRICS_LABEL])
+        .inc_by(total);
+    PROVER_NUM_METRICS
+        .with_label_values(&[TOTAL_METRIC_FAMILIES_OVER_2000_LABEL])
         .inc_by(families_over_2000);
 
     metric_families
